data/hil-orchestration/hil-agent-system/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & Caching
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.23
alembic>=1.13.0
redis>=5.0.1
psycopg2-binary>=2.9.9

# Vector Database
pgvector>=0.2.4

# LLM Providers
openai>=1.3.0
anthropic>=0.7.0

# Tool Integration
composio-core>=0.3.0
composio-openai>=0.3.0

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Templating & Data
jinja2>=3.1.2
pyyaml>=6.0.1
python-multipart>=0.0.6

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Observability
structlog>=23.2.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0

# Workflow & Task Management
celery>=5.3.4
networkx>=3.2.1

# Utilities
tenacity>=8.2.3
click>=8.1.7
rich>=13.7.0
typer>=0.9.0